requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.4
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==19.9.0
importlib-metadata==0.20
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==3.0.1
mongoengine==0.18.2
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
python-dotenv==0.10.3
six==1.12.0
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
